Automatically start embedded yaws-server
In rebar3-shell the yaws is now started automatically.

The README.md was updated - removing the explanation
of the obj_fetch/3 method in yaws json2-module.

rebar3-deps now point to original yaws-repo instead of
patched fnchooft/yaws-repo after incorporating patch.

The ybed_sup.erl testcode was removed and incorporated
into the yaws_rest_sup, and ybed.erl was updated to
instantiate below the correct supervisor.

The route to the airport-rest endpoint is now:
"/rest/airport" instead of "/".